w3c-reports/12092014/mac/runned-tests/html/semantics/embedded-content/media-elements/interfaces/TextTrackCue/startTime-expected.txt

CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.


PASS TextTrackCue.startTime, script-created cue 
FAIL TextTrackCue.startTime, parsed cue assert_unreached: got error event Reached unreachable code(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_unreached@http://localhost:8800/resources/testharness.js:692:16
http://localhost:8800/html/semantics/embedded-content/media-elements/interfaces/TextTrackCue/startTime.html:32:23
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
